requirements/requirements_lumi_pi4.txt
absl-py==0.9.0
appdirs==1.4.3
asn1crypto==0.24.0
astor==0.8.1
astroid==2.1.0
asttokens==1.1.13
automationhat==0.2.0
beautifulsoup4==4.7.1
bleach==1.5.0
blinker==1.4
blinkt==0.1.2
buttonshim==0.0.2
cachetools==4.1.1
Cap1xxx==0.1.3
certifi==2018.8.24
chardet==3.0.4
click==6.7
colorama==0.3.7
colorzero==1.1
cookies==2.2.1
cryptography==2.6.1
cycler==0.10.0
Cython==0.29.21
decorator==4.3.0
dm-sonnet==1.23
docutils==0.14
drumhat==0.1.0
easydict==1.8
entrypoints==0.3
envirophat==1.0.0
ExplorerHAT==0.4.2
Flask==1.0.2
fourletterphat==0.1.0
gast==0.2.2
google-auth==1.20.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
gpiozero==1.5.1
grpcio==1.30.0
guizero==0.6.0
h5py==2.9.0
html5lib==0.9999999
idna==2.6
importlib-metadata==1.7.0
ipykernel==4.9.0
ipython==7.16.3
ipython-genutils==0.2.0
isort==4.3.4
itsdangerous==0.24
jedi==0.13.2
Jinja2==2.10
joblib==0.14.1
jupyter-client==5.2.3
jupyter-core==4.4.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==17.1.1
keyrings.alt==3.1.1
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
logilab-common==1.4.2
-e git+git@github.com:czbiohub/luminoth-uv-imaging.git@master#egg=luminoth
lxml==4.6.5
Markdown==3.2.2
MarkupSafe==1.1.0
matplotlib==3.0.2
mccabe==0.6.1
microdotphat==0.2.1
mote==0.0.4
motephat==0.0.3
mypy==0.670
mypy-extensions==0.4.1
natsort==6.2.0
nudatus==0.0.4
numpy==1.21.0
oauthlib==2.1.0
olefile==0.46
opencv-contrib-python==4.2.0.32
opencv-python==4.2.0.32
opt-einsum==3.3.0
pandas==1.1.0rc0
pantilthat==0.0.7
parso==0.3.1
pexpect==4.6.0
pgzero==1.2
phatbeat==0.1.1
pianohat==0.1.0
picamera==1.13
pickleshare==0.7.5
picraft==1.0
piglow==1.2.5
pigpio==1.44
Pillow==9.0.0
prompt-toolkit==1.0.15
protobuf==3.12.4
psutil==5.5.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.4.0
pycrypto==2.6.1
pyflakes==2.0.0
pygame==1.9.4.post1
Pygments==2.7.4
PyGObject==3.30.4
pyinotify==0.9.6
PyJWT==1.7.0
pylint==2.2.2
pyOpenSSL==19.0.0
pyparsing==2.2.0
pyserial==3.4
python-apt==1.8.4.1
python-dateutil==2.7.3
pytz==2020.1
pyxdg==0.25
PyYAML==5.4
pyzmq==17.1.2
qtconsole==4.3.1
rainbowhat==0.1.0
requests==2.18.4
requests-oauthlib==1.0.0
responses==0.9.0
roman==2.0.0
RPi.GPIO==0.7.0
rsa==4.7
RTIMULib==7.2.1
scikit-learn==0.19.2
scikit-video==1.1.10
scipy==1.1.0
scrollphat==0.0.7
scrollphathd==1.2.1
SecretStorage==2.3.1
semver==2.0.1
Send2Trash==1.5.0
sense-emu==1.1
sense-hat==2.2.0
simplegeneric==0.8.1
simplejson==3.16.0
six==1.12.0
skywriter==0.0.7
sn3218==1.2.7
soupsieve==1.8
spidev==3.4
ssh-import-id==5.7
tensorboard==1.13.1
tensorflow==2.5.3
tensorflow-estimator==1.14.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
thonny==3.2.6
threadpoolctl==2.1.0
tornado==5.1.1
touchphat==0.0.1
traitlets==4.3.2
twython==3.7.0
typed-ast==1.3.1
uflash==1.2.4
unicornhathd==0.0.4
urllib3==1.26.5
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
wrapt==1.12.1
XlsxWriter==1.3.0
zipp==3.1.0
